Changed the UUID characteristic to UUIDString
I don't know if the UUID of the characteristic was intentionally put to be its description. Currently, the description value is displayed (which can be the UUIDString or even the feature characteristic - in my project, is displayed the name), which is quite difficult when you need the characteristic ID and not its name.